% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{age_calc}
\alias{age_calc}
\title{Given a list of IDates, compute associated ages.}
\usage{
age_calc(vdate0, date1 = as.Date(Sys.time()))
}
\arguments{
\item{vdate0}{\code{IDate} vector}

\item{date1}{\code{IDate} vector of length one, or of the same length as \code{vdate0}}
}
\description{
Given a list of IDates, compute associated ages.
}
