% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{countby}
\alias{countby}
\title{Tabulation utility for data.table}
\usage{
countby(
  DT,
  .x,
  bycols,
  .x_excludena = FALSE,
  bycols_excludena = FALSE,
  wide = TRUE,
  .w = NULL,
  side_effects = FALSE
)
}
\description{
Tabulate columns specified in \code{.x} of dataset \code{DT} (of class \code{data.table}),
groupped by columns specified in \code{bycols}. Allows specifying how missing
values should be handled both in grouping and the target variables.
}
\details{
Allows simple weights, specified by \code{.w}. When weights are provided, the
element count is the sum of weights in the group for a particular value.
}
