% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{.lunique}
\alias{.lunique}
\title{List unique values of \code{data.table} object columns}
\usage{
.lunique(DT, z, exclude_na = NULL)
}
\arguments{
\item{DT}{the data.table object}

\item{z}{the column names whose unique values are to be computed}

\item{exclude_na}{either a logical value or a subset of \code{z} values}
}
\description{
Compute a named list whose element \code{i} corresponds to the unique values of
the data.table \code{DT} column \code{z[i]}. If \code{exclude_na} is \code{NULL} or \code{FALSE},
missing values are included in the list of unique values. If \code{exclude_na = TRUE}, missing values are excluded for all columns in \code{z}. If \code{exclude_na} is
a subset of \code{z}, the missing values of those columns will be excluded.
}
