% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{.wsummby}
\alias{.wsummby}
\title{Weighted summary  by category (internal function)}
\usage{
.wsummby(DT, .w, .x, bycols, bycols_excludena = NULL, f, ...)
}
\description{
Weighted summary  by category (internal function)
}
