% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{summby}
\alias{summby}
\title{Weighted summary by category}
\usage{
summby(DT, .x, bycols, f, bycols_excludena = NULL, ..., .w = NULL)
}
\arguments{
\item{DT}{the \code{data.table} data set}

\item{.x}{the name of the column to be summarized}

\item{bycols}{the name of the columns by which we are computing the summaries}

\item{bycols_excludena}{one of \code{NULL}, \code{TRUE}, \code{FALSE} or a character vector
specifying which of \code{bycols} should have NAs ignored}

\item{.w}{lenth 1 character vector indicating weight column}
}
\value{
a \code{data.table} with the aggregation
}
\description{
This is a wrapper around \code{data.table} aggregation by category that can be
reused in different contexts with user supplied column names, and specific
behavior about NA categories.
}
