% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.r
\name{union_ids}
\alias{union_ids}
\title{Create a maximal ID based on 2 id'ing schemes}
\usage{
union_ids(DT, c1, c2, name_join, maxiter = 50)
}
\arguments{
\item{DT}{the data.table object}

\item{c1}{a string with the column name for the first id}

\item{c2}{a string with the column name for the second id}

\item{exclude_na}{either a logical value or a subset of \code{z} values}
}
\description{
Create a maximal ID based on 2 id'ing schemes
}
